+++ src/sll/vm.c
@@ -5,6 +5,7 @@
 #include <sll/constants.h>
 #include <sll/core.h>
 #include <sll/gc.h>
+#include <sll/handle.h>
 #include <sll/map.h>
 #include <sll/operator.h>
 #include <sll/static_object.h>
@@ -64,6 +65,7 @@ const sll_runtime_data_t* sll_current_runtime_data=NULL;
 __SLL_FUNC __SLL_RETURN_CODE sll_execute_assembly(const sll_assembly_data_t* a_dt,const sll_stack_data_t* st,const sll_runtime_data_t* r_dt,sll_error_t* e){
 	e->t=SLL_ERROR_UNKNOWN;
 	sll_current_runtime_data=r_dt;
+	sll_handle_type_t hl_l=r_dt->hl->dtl;
 	sll_internal_function_table_t* i_ft=r_dt->ift;
 	sll_output_data_stream_t* out=r_dt->os;
 	sll_return_code_t rc=0;
@@ -637,6 +639,7 @@ _return:;
 		}
 	}
 _end:
+	sll_cleanup_handles(r_dt->hl,hl_l);
 	if (e->t==SLL_ERROR_UNKNOWN){
 		if (!sll_verify_runtime_object_stack_cleanup(&rst)){
 			e->t=SLL_ERROR_UNRELEASED_OBJECTS;